Change new function names to use `_` instead of `.`
Occasionally, R functions are intepreted to be object-oriented methods (I can explain more next time we meet) when they include a `.`. It looks like some of your [x].to.speed functions might have been doing this. This wasn't the error in the build, but it will be cleaner if we make them `_` instead (I would change the earlier functions, too, but it's probably best not to in case anyone's using them in their own code-- it would break their code when the new version came out until they went through and changed all their function names.)