Close #13
If you just give part of an argument name (`r` here), R will try to match an argument that starts that way. It's find if you only have one argument that starts with "r", but here you have two ("round" and "round.out"), so R isn't sure which you're refering to. Spelling out the argument name fixed this.